Fix issue where IBD of large blocks was disconnecting nodes during download.  Fix issue where net/recv/total only is incremented upon receipt of an entire message.  Add RequestManager::RemoveSource() function -- a quick abort of ongoing requests when the relevant node disconnects

+++ src/globals.cpp
@@ -113,7 +113,7 @@ set<CNetAddr> setservAddNodeAddresses;
 uint64_t maxGeneratedBlock = DEFAULT_MAX_GENERATED_BLOCK_SIZE;
 uint64_t excessiveBlockSize = DEFAULT_EXCESSIVE_BLOCK_SIZE;
 unsigned int excessiveAcceptDepth = DEFAULT_EXCESSIVE_ACCEPT_DEPTH;
-unsigned int maxMessageSizeMultiplier = DEFAULT_MAX_MESSAGE_SIZE_MULTIPLIER;
+uint64_t maxMessageSizeMultiplier = DEFAULT_MAX_MESSAGE_SIZE_MULTIPLIER;
 int nMaxOutConnections = DEFAULT_MAX_OUTBOUND_CONNECTIONS;
 
 uint32_t blockVersion = 0; // Overrides the mined block version if non-zero
